Bump down to 9600 baud to work around XON/XOFF
The ATWINC1500 seems to have a firmware bug that makes a write()
hang sometimes if there's incoming data to handle.  The Arduino library
doesn't seem to work around this yet, but bumping down the speed
means we're less like to send XOFF/XON on an active connection
because the terminal can keep up (putting chars on the screen).